Restructure new reporting functionality into its own app
