+++ rdrf/explorer/static/css/explorer.css
@@ -7,10 +7,6 @@ body {
     margin-bottom: 30px;
 }
 
-.form .row .col-form-label {
-    text-align: left;
-}
-
 #footer {
     position: absolute;
     bottom: 0;
