+++ rdrf/explorer/urls.py
@@ -1,7 +1,7 @@
 from django.urls import re_path
 
-from explorer.views import DeleteQueryView, DownloadQueryView, ReportDeleteView, ReportDownloadView
-from explorer.views import MainView, ReportsView, ReportDesignView
+from explorer.views import DeleteQueryView, DownloadQueryView
+from explorer.views import MainView
 from explorer.views import QueryView, NewQueryView
 from explorer.views import SqlQueryView
 
@@ -17,13 +17,6 @@ urlpatterns = [
 
     re_path(r'^sql$',
             SqlQueryView.as_view(), name='explorer_sql_query'),
-
-    re_path(r'^report$', ReportDesignView.as_view(), name='explorer_report_designer'),
-    re_path(r'^report/(?P<report_id>\w+)/?$', ReportDesignView.as_view(), name='explorer_report_designer'),
-    re_path(r'^report/(?P<report_id>\w+)/delete/?$', ReportDeleteView.as_view(), name='explorer_report_delete'),
-    re_path(r'^report/download/(?P<report_id>\w+)/(?P<format>\w+)/?$', ReportDownloadView.as_view(),
-            name='explorer_report_download'),
-    re_path(r'^reports$', ReportsView.as_view(), name='explorer_reports_list'),
 
     re_path(r'^new$', NewQueryView.as_view(), name='explorer_new'),
