+++ rdrf/rdrf/security/url_whitelist.py
@@ -349,9 +349,9 @@ SECURITY_WHITELISTED_URLS = (
     "rdrf:explorer_query_delete",
     "rdrf:explorer_query_download",
     "rdrf:explorer_sql_query",
-    "rdrf:explorer_report_designer",
-    "rdrf:explorer_report_download",
-    "rdrf:explorer_reports_list",
+    "report:report_designer",
+    "report:report_download",
+    "report:reports_list",
     "registration_activate",
     "registration_activation_complete",
     "registration_complete",
