+++ rdrf/report/static/css/report.css
@@ -1,0 +1,3 @@
+.form .row .col-form-label {
+    text-align: left;
+}
