+++ rdrf/report/templates/base.html
@@ -20,7 +20,7 @@
         <link rel="stylesheet" href="{% static 'css/vendor/font-awesome.min.css' %}">
 
         <link rel="stylesheet" type="text/css" href="{% static 'css/rdrf.css' %}" media="all">
-        <link rel="stylesheet" href="{% static 'css/explorer.css' %}">
+        <link rel="stylesheet" href="{% static 'css/report.css' %}">
         <script type="text/javascript" src="{% static 'js/rpc_module.js' %}"></script>
         
         <script type="text/javascript" src="{% static 'js/jquery.cookies.js' %}"></script>
